+++ src/cmysql/c_mysql_handshake_packet.h
@@ -12,26 +12,30 @@
 #include <iostream>
 using namespace std;
 
+#include <string.h>
+#include <stdint.h>
+
 class CMysqlHandshakePacket {
 public:
 	CMysqlHandshakePacket();
 	virtual ~CMysqlHandshakePacket();
 
 public:
-	void set_thread_id(int id);
-	int serialize(char *buffer,int len,long int &pos);
+	void set_thread_id(uint32_t id);
+	int serialize(char *buffer,uint64_t len,uint64_t &pos);
 
 private:
     uint8_t protocol_version_;
     string server_version_;
-    int thread_id_;//connection id
+    uint32_t thread_id_;//connection id
     char scramble_buff_[8];// auth_plugin_data_part_1 : first 8 bytes of the auth-plugin data
-    int server_capabilities_;  /* set value to use 4.1protocol */
-    int server_language_;
-    int server_status_;
+    uint16_t server_capabilities_;  /* set value to use 4.1protocol */
+    uint8_t server_language_;
+    uint16_t server_status_;
     char plugin_[13];        /* always 0x00 */
     char plugin2_[12];
     char terminated_;
+
     string str_buf_;   //store ObString content
